Add tests for required props utils, share setup with getAllProps tests
